requirements.txt
tkinter
python-statemachine
pyperclip
pynput
google-api-python-client
google-cloud-speech
pyaudio
termcolor